Refactored: Defined new redux patterns for better organization and maintenance
