--- src/redux/actions.ts
@@ -1,1 +1,0 @@
-export { setState } from "redux/slices/buy-dag";
