+++ src/redux/states/buyDag/buyDag.actions.ts
@@ -1,0 +1,19 @@
+
+/////////////////////////////
+// ANCHOR Actions
+/////////////////////////////
+
+function setPaymentInfo(state, { payload }) {
+  return {
+    ...state,
+    ...payload,
+  };
+}
+
+/////////////////////////////
+// ANCHOR Exports
+/////////////////////////////
+
+export default {
+  setPaymentInfo,
+}
