+++ src/redux/states/buyDag/index.ts
@@ -1,0 +1,2 @@
+export { actions } from "./buyDag.slice";
+export { reducer } from "./buyDag.slice";
